refactor(repo): split backup_export.go into proxy/security/settings files
